chore: simplify native post turbo modules

+++ android/src/oldarch/io/customer/reactnative/sdk/logging/NativeCustomerIOLoggingModule.kt
@@ -44,7 +44,7 @@ class NativeCustomerIOLoggingModule(
     }
 
     @ReactMethod
-    fun removeListeners(count: Int) {
-        listenerCount -= count
+    fun removeListeners(count: Double) {
+        listenerCount -= count.toInt()
     }
 }
